Move beehive tests to beehive directory
